x86_64: Introduce a new module to manage MSRs
Remove the currently scattered definitions of MSRs and merge it into one
module.
